Reorganize codebase and remove unused code
