+++ smart-contracts/test/VotingTest.js
@@ -1,4 +1,4 @@
-const catchRevert = require('../helper/exception').catchRevert;
+const catchRevert = require('./helper/exception').catchRevert;
 const Voting = artifacts.require('./Voting');
