+++ smart-contracts/test/constant/index.js
@@ -1,0 +1,10 @@
+module.exports.Membership = {
+    NOBODY: 0,
+    CITIZEN: 1,
+    DIAMOND: 2
+}
+
+module.exports.MembershipPricing = {
+    CITIZEN: 1,
+    DIAMOND: 10
+}
