+++ smart-contracts/test/helper/exception.js
